[BREAKING CHANGE] Change how claims are handled (#221)
This is a breaking change. Please read the updated docs and the PR description thoroughly.

The major changes are:

Claims validation is now provided as a function when configuring the handler
Claims are typed based on the provided type (generics are used)
options.WithRequiredClaims() is deprecated
Now, configuring a handler requires you to define what type the claims are. You can still use map[string]interface{} if you want and the parameter for ClaimsValidationFn can be set to nil.